Chrome browser extension to execute background NDT7 Wi-Fi tests
Extension requires a Netrics device attached to the network.

As such, extension "sniffs" out the address of this device -- currently,
only checking for it at either "netrics.local" or "netrics.localdomain".

Wi-Fi testing is negotiated with and reported to the Dashboard Trial API
(as in the dashboard page).

The extension's toolbar icon button serves to open the Dashboard --
either any already-open tab or in a new tab -- and as such serves to aid
participants in navigating to the Dashboard (whose address cannot be
assumed).

Upon installation, the extension attempts to discover the Netrics device
adddress and open the Dashboard right away -- introducing the
participant to the Dashboard.